`s.any` now conforms values using the first successful Spec's conformer (#39)
Specs created using the `s.any` factory will now use the conformer of the first successful constituent Spec as the default conformer. Callers may specify an outer conformer which will be applied to the conformed value. Additionally, callers may specify `tag_conformed` to emit the final value wrapped in a 2-tuple, where the first element is the tag of the successfully validating Spec and the second element is the conformed value.