Key/value Specs (#76)
This commit adds support for "generic" mapping specs over homogeneous key/value mappings. Unlike the base constructor `s({k: v, ...})`, which produces Specs for maps with known keys, `s.kv(...)` is for maps where every key/value matches a known Spec, but for which the keys may not be known a priori.