Only include user-defined (or default) tags in ErrorDetails.via (#81)
Before this commit, Spec factories would convert all of their individual `ValidatorFn`s into `ValidatorSpecs` and compose the Specs together. This meant that each failing `ValidatorFn` in a factory-created Spec would append its own child tags into the `ErrorDetails.via` if a value violated that specific sub-Spec. Originally, I thought this may be useful to know which particular factory check failed ("this value didn't meet the regex!"), but after working through several different implementations we found that it was too difficult to determine just which Spec failed since the actual failing Spec from the perspective of the developer using `dataspec` is the one they tagged.

This change converts `ValidatorSpec.from_validators` to a classmethod that simply composes multiple `ValidatorFn`s, rather than composing them into `ValidatorSpecs` and then wrapping that.

The only place this created a slight challenge was with the `s.str(...)` formatting code, which was actually relying on the fact that it could send a `ValidatorSpec` directly (to allow for a conformation stage). I think I managed to work around this though by just composing the conformer directly in the `s.str(...)` if using a conforming format.

I've also converted all `Optional[Tag]` parameters to just `Tag` and supplied the default value directly in the signature to surface factory tags to users directly (in addition to adding the tag to the docstring for the parameter). My hope is this will make it clearer to users exactly what tags will be provided if they do not give one.

All of the composing-type factories (`s.all`, `s.any`, `s.blankable`, `s.default`, `s.merge`, and `s.nilable`) have proper validation that they are receiving only one spec/spec-coercible value. Additionally, `s.blankable` and `s.nilable` should now emit slightly more useful tags in `ErrorDetails.via` since the failure to validate the `None`/`""` value should match the `nilable`/`blankable` tag, whereas the failure of the inner Spec should match that tag.